requirements.txt
# Core deep learning for CUDA 12.1
--extra-index-url https://download.pytorch.org/whl/cu121
 
# Hugging Face stack (compatible with torch 2.2.2)
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
xformers==0.0.24 # Requires torch==2.2.0, but often works with 2.2.x

# Face detection & preprocessing
facenet-pytorch==2.6.0 # Requires numpy<2.0.0
opencv-python-headless==4.8.0.76 # This version is compatible with numpy < 2.0.0
numpy<2.0.0 # Explicitly set numpy version to satisfy facenet-pytorch

# Utils
pillow
tqdm
matplotlib
scikit-learn
seaborn
scikit-image

# Config & logging
pyyaml